\name{betaABfromMeanKappa}
\alias{betaABfromMeanKappa}
\title{
Beta parameters from mean and concentration
}
\description{
Returns the parameters (a, b) of a beta distribution from estimates of the mean and
concentration
}
\usage{
betaABfromMeanKappa(mean, kappa)
}
\arguments{
\item{mean}{scalar between 0 and 1}
\item{kappa}{scalar greater than 0}
}
\value{
Invisibly a named list with parameters a and b
}
\author{
\cr
David Brown\cr
\email{brownd7@mskcc.org}\cr
}
\keyword{}
