\name{betaABfromMeanSD}
\alias{betaABfromMeanSD}
\title{
Beta parameters from mean and standard deviation
}
\description{
Returns the parameters (a, b) of a beta distribution from estimates of the mean and
standard deviation
}
\usage{
betaABfromMeanSD(mean, sd)
}
\arguments{
\item{mean}{scalar between 0 and 1}
\item{sd}{scalar greater than 0}
}
\value{
Invisibly a named list with parameters a and b
}
\author{
\cr
David Brown\cr
\email{brownd7@mskcc.org}\cr
}
\keyword{}
