\name{consensusBaseatCpG}
\alias{consensusBaseatCpG}
\title{
Computes the consensus for paired-end sequences where the two ends overlap
}
\description{
Computes the consensus for paired-end sequences at positions where the two ends overlap
}
\usage{
consensusBaseatCpG(alignment)
}
\arguments{
\item{alignment}{list containing a GAlignmentPairs metadata of the sequences and two data frames of sequences and base qualities}
}
\value{
Invisibly a list containing a GAlignmentPairs metadata of the sequences and two data frames of sequences and base qualities
}
\author{
\cr
David Brown\cr
\email{brownd7@mskcc.org}\cr
}
\keyword{}
