\name{package:EMseq}
\docType{package}
\alias{EMseq}
\title{
	Enzymatic Methylation sequencing
}
\description{
Read level detection of cancer from bisulfite-free methylation sequencing
}
\details{
	\tabular{ll}{
		Package: \tab zzz\cr
		Type: \tab Package\cr
		Version: \tab 0.0-1\cr
		Date: \tab 2020-11-08\cr
		License: \tab GPL (>= 2)\cr
		LazyLoad: \tab yes\cr
	}
}
\author{
	\tabular{ll}{
		author(s): \tab David Brown\cr
				   \tab brownd7@mskcc.org\cr
				   \tab \cr
		maintainer(s): \tab David Brown\cr
					   \tab brownd7@mskcc.org\cr
	}
}
\keyword{package}