\name{gammaShRaFromModeSD}
\alias{gammaShRaFromModeSD}
\title{
Gamma parameters from mode and standard deviation
}
\description{
Returns the shape and rate parameters of a gamma distribution from estimates of the mode
and standard deviation
}
\usage{
gammaShRaFromModeSD(mode, sd)
}
\arguments{
\item{mode}{scalar greater than 0}
\item{sd}{scalar greater than 0}
}
\value{
Invisibly a named list with parameters shape and rate
}
\author{
\cr
David Brown\cr
\email{brownd7@mskcc.org}\cr
}
\keyword{}
