\name{LL}
\alias{LL}
\title{
Likelihood
}
\description{
Returns the likelihood
}
\usage{
LL(p, n, p_hat, a_hat, b_hat, log = TRUE)
}
\arguments{
\item{p}{}
\item{n}{}
\item{p_hat}{}
\item{a_hat}{}
\item{b_hat}{}
\item{log}{}
}
\value{
Invisibly the value of the likelihood
}
\author{
\cr
David Brown\cr
\email{brownd7@mskcc.org}\cr
}
\keyword{}
