\name{consensusBaseatCpG}
\alias{consensusBaseatCpG}
\title{
...
}
\description{
...
}
\usage{
consensusBaseatCpG(pileup)
}
\arguments{
\item{alignment}{list containing a GAlignmentPairs metadata of the sequences and two data frames of sequences and base qualities}
}
\value{
Invisibly a list containing a GAlignmentPairs metadata of the sequences and two data frames of sequences and base qualities
}
\author{
\cr
David Brown\cr
\email{brownd7@mskcc.org}\cr
}
\keyword{}
